ENSO/app0-admin/requirements.txt
hopeit.engine[web,cli,fs-storage,redis-streams,redis-storage,apps-visualizer,config-manager,apps-client]>=0.19.0,<0.20.0
motor[srv]>=2.5.1
jinja2>=3.1.1
aioboto3>=9.6.0
bcrypt>=4.0.1
aiosmtplib>=2.0.2
pendulum>=2.1.2
